feat: move zip logic to compression service
