+++ src/Core/VideoProcessor.Application/DependencyInjection/ServiceCollectionExtensions.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace VideoProcessor.Application.DependencyInjection
+{
+    public static class ServiceCollectionExtensions
+    {
+        public static IServiceCollection AddUseCases(this IServiceCollection services)
+        {
+
+            return services;
+        }
+    }
+}
