+++ src/Core/VideoProcessor.Application/Services/ICompressionService.cs
@@ -1,0 +1,9 @@
+﻿using VideoProcessor.Domain.Entities;
+
+namespace VideoProcessor.Application.Services
+{
+    public interface ICompressionService
+    {
+        BinaryFile CreateZipFile(IEnumerable<BinaryFile> files, string outputZipFileName)
+    }
+}
